+++ srcs/backend/src/server.ts
@@ -16,10 +16,7 @@ import fastifyPassport from '@fastify/passport'
 import fastifySecureSession from '@fastify/secure-session'
 import { VerifyCallback } from 'passport-google-oauth2'
 import { registerSchemas } from './schemas/list';
-import type { RawData, WebSocket as WSSocket } from 'ws';
-import { joinRoom, leaveRoom, leaveAll, broadcast } from './rooms';
-import { enqueue } from './onlineTourn';
-import { createTournament,getByCode, tournaments } from './realtimeTournament';
+import { wsRoutes } from './routes/ws.routes';
 
 var GoogleStrategy = require('passport-google-oauth2').Strategy;
 
@@ -80,6 +77,7 @@ const startServer = async () => {
 		await app.register(authRoutes);
 		await app.register(gameRoutes);
 		await app.register(friendRoutes);
+		await app.register(wsRoutes);
 
 		fastifyPassport.use('google', new GoogleStrategy({
 			clientID:     process.env.GOOGLE_CLIENT_ID,
@@ -100,104 +98,6 @@ const startServer = async () => {
 			};
 			}
 		));
-		app.get('/ws', { websocket: true }, (conn, req) => {
-			// compatible with both handler shapes
-			const raw = conn as any;
-			const socket: WSSocket = (raw && raw.socket) ? (raw.socket as WSSocket) : (raw as WSSocket);
-			
-			let myTournamentId: string | null = null;
-			let myPlayerId: string | null = null;
-			// greet
-			setTimeout(() => {
-			  if (socket.readyState === socket.OPEN) {
-				socket.send(JSON.stringify({ type: 'hello', ts: Date.now() }));
-			  }
-			}, 0);
-		  
-			socket.on('message', (data: RawData) => {
-			  const text = typeof data === 'string' ? data : data.toString();
-			  let msg: any; try { msg = JSON.parse(text); } catch { return; }
-		  
-			  switch (msg?.type) {
-				case 'ping':
-				  socket.send(JSON.stringify({ type: 'pong', ts: Date.now() }));
-				  break;
-		  
-				case 'queue':
-				  // enqueue this socket; pass goalLimit if you want it customizable
-				  enqueue(socket, typeof msg.goalLimit === 'number' ? msg.goalLimit : 5);
-				  // (optional) ack
-				  socket.send(JSON.stringify({ type: 'queued' }));
-				  break;
-				  case 't.create': {
-					const name = String(msg.name ?? 'Tournament');
-					const goalLimit = Number(msg.goalLimit ?? 5);
-					const maxPlayers = Math.min(8, Math.max(2, Number(msg.maxPlayers ?? 8)));
-					const alias = String(msg.alias ?? 'Host');
-			
-					const t = createTournament({ name, goalLimit, maxPlayers });
-					const pid = t.addPlayer(socket, alias);
-					t.setHost(pid);
-			
-					myTournamentId = t.id;
-					myPlayerId = pid;
-			
-					socket.send(JSON.stringify({ type: 't.created', id: t.id, code: t.code }));
-					// t.addPlayer already broadcasted t.state to lobby
-					break;
-				  }
-			
-				  // Anyone joins by code
-				  case 't.join': {
-					const code = String(msg.code ?? '');
-					const alias = String(msg.alias ?? 'Player');
-					const t = getByCode(code);
-					if (!t) {
-					  socket.send(JSON.stringify({ type: 't.error', msg: 'Invalid code' }));
-					  break;
-					}
-					const pid = t.addPlayer(socket, alias);
-					myTournamentId = t.id;
-					myPlayerId = pid;
-			
-					socket.send(JSON.stringify({ type: 't.joined', id: t.id, pid }));
-					// t.addPlayer already broadcasted t.state
-					break;
-				  }
-			
-				  // Player toggles ready
-				  case 't.ready': {
-					if (!myTournamentId || !myPlayerId) break;
-					const t = tournaments.get(myTournamentId);
-					if (!t) break;
-					t.markReady(myPlayerId, !!msg.ready);
-					break;
-				  }
-			
-				  // Host starts tournament
-				  case 't.start': {
-					if (!myTournamentId || !myPlayerId) break;
-					const t = tournaments.get(myTournamentId);
-					if (!t) break;
-					if (t.host !== myPlayerId) {
-					  socket.send(JSON.stringify({ type: 't.error', msg: 'Only host can start' }));
-					  break;
-					}
-					t.start();
-					break;
-				  }
-			
-				  default:
-					// Ignore unknown; match inputs are handled by the match loop
-					break;
-				}
-			  });
-		  
-			const hb = setInterval(() => {
-			  if (socket.readyState === socket.OPEN) socket.ping();
-			}, 30_000);
-			socket.on('close', () => clearInterval(hb));
-		  });
 		await app.listen({ port: 3000, host: '0.0.0.0' });
 
 		console.log('Server running at http://localhost:3000')
